refactor: merge client files and remove unused config
- Merged client.py and client_sync.py into a single file with both async and sync examples
- Added interactive mode selection for users
- Removed unused mcp_servers_config.json file
- Addresses PR review feedback from MuggleJinx
