+++ examples/toolkits/mcp/mcp_terminal_toolkit/client.py
@@ -19,7 +19,8 @@ from mcp.types import CallToolResult
 from camel.toolkits.mcp_toolkit import MCPClient, MCPToolkit
 
 
-async def run_example_stdio():
+async def run_example_async():
+    """Example of using Terminal Toolkit MCP server asynchronously."""
     config = {
         "mcpServers": {
             "terminal_toolkit": {
@@ -48,5 +49,49 @@ async def run_example_stdio():
         print(res1.content[0].text[:1000])
 
 
+def run_example_sync():
+    """Example of using Terminal Toolkit MCP server synchronously."""
+    config = {
+        "mcpServers": {
+            "terminal_toolkit": {
+                "command": sys.executable,
+                "args": [
+                    "examples/toolkits/mcp/mcp_terminal_toolkit/terminal_toolkit_server.py",
+                    "--mode",
+                    "stdio",
+                ],
+            }
+        }
+    }
+    mcp_toolkit = MCPToolkit(config_dict=config, timeout=60)
+    with mcp_toolkit:
+        mcp_client: MCPClient = mcp_toolkit.clients[0]
+        res = mcp_client.list_mcp_tools_sync()
+        if isinstance(res, str):
+            raise Exception(res)
+
+        tools = [tool.name for tool in res.tools]
+        print(f"Available tools: {tools}")
+
+        res1: CallToolResult = mcp_client.call_tool_sync(
+            "shell_exec", {"id": "test_session_sync", "command": "ls -a"}
+        )
+        print(res1.content[0].text[:1000])
+
+
 if __name__ == "__main__":
-    asyncio.run(run_example_stdio())
+    print("Choose mode:")
+    print("1. Async example")
+    print("2. Sync example")
+    
+    choice = input("Enter your choice (1 or 2): ").strip()
+    
+    if choice == "1":
+        print("Running async example...")
+        asyncio.run(run_example_async())
+    elif choice == "2":
+        print("Running sync example...")
+        run_example_sync()
+    else:
+        print("Invalid choice. Running async example by default...")
+        asyncio.run(run_example_async())
