--- examples/toolkits/mcp/mcp_terminal_toolkit/mcp_servers_config.json
@@ -1,12 +1,0 @@
-{
-  "mcpServers": {
-    "terminal_toolkit": {
-      "command": "python",
-      "args": [
-        "examples/toolkits/mcp/mcp_terminal_toolkit/terminal_toolkit_server.py",
-        "--mode",
-        "stdio"
-      ]
-    }
-  }
-}
